FizzBuzz v0.1.0 FizzBuzz

A module that contains fizz-buzz helpers.

What’s fizz-buzz?

It’s a simple programming task that checks the base programming knowledge. It takes a number (or a range) and it should print/return the following:

“Fizz” for values divisible by 3,

“Buzz” for values divisible by 5,

“FizzBuzz” for values divisible by both 3 and 5

the passed value if none of the above matches

Functions

get_range(start, finish)

Gets a list of FizzBuzz values from a range

iex> FizzBuzz.get_range(0, 15)
[0, 1, 2, "Fizz", 4, "Buzz", "Fizz", 7, 8, "Fizz", "Buzz", 11, "Fizz", 13, 14, "FizzBuzz"]

get_single_value(number)

Gets a single FizzBuzz value from the argument

iex> FizzBuzz.get_single_value(15)
"FizzBuzz"

print_range(start, finish)

Prints a range of FizzBuzz cases, from start to finish, using the FizzBuzz.get_range function